+++ sumologic/resource_sumologic_cse_network_block_test.go
@@ -6,8 +6,8 @@ import (
 	"net"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 func TestAccSumologicSCENetworkBlock_create(t *testing.T) {